完成锁机和解锁的代码，并做过简单测试。 1. 锁机: 创建一个窗口占满全屏，限制键盘和鼠标操作 解锁: 做相反的操作。

+++ RemoteCtrl/RemoteCtrl/ServerSocket.h
@@ -36,7 +36,7 @@ public:
 		: head(0), length(0), cmd(0), sum(0)
 	{
 		head = PACK_HEAD;
-		length = size + 4;
+		length = (DWORD)(size + 4);
 		cmd = _cmd;
 		if (size > 0)
 		{
@@ -98,7 +98,7 @@ public:
 	bool SendData(const char* data, size_t size)
 	{
 		if (m_client == INVALID_SOCKET) return false;
-		return send(m_client, data, size, 0) > 0;
+		return send(m_client, data, (int)size, 0) > 0;
 	}
 
 	bool SendData(CPacket& packet)
